Migrate box Plot (Deprecated) visualization to React
