+++ client/app/pages/dashboards/dashboard.less
@@ -84,6 +84,7 @@
     .sankey-visualization-container,
     .map-visualization-container,
     .word-cloud-visualization-container,
+    .box-plot-deprecated-visualization-container,
     .plotly-chart-container {
       position: absolute;
       left: 0;
