+++ client/app/visualizations/box-plot/renderer.less
@@ -1,0 +1,4 @@
+.box-plot-deprecated-visualization-container {
+  overflow: hidden;
+  height: 400px;
+}
